Reorganize + Test Implement Signup
 - Reorganized the file structure of the SignUp Component so that the types
   are sub-routes.
 - Moved the two types of signup forms to a forms folder.
 - Routes are currently test routes for demonstration purposes. This is
   not a complete feature.
