+++ client/src/components/SignUp/Types/SignUpUserTypes.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import { ModalContainer } from '../../styles/ModalGlobalStyle.js';
-import { Label, Input, SignModalMain } from '../../styles/SignIn_UpStyle';
 
 class SignUpUserTypes extends React.Component {
   state = {
@@ -23,15 +21,11 @@ class SignUpUserTypes extends React.Component {
   render() {
     // console.log('%cstate', 'color: blue', this.state);
     return (
-      <ModalContainer data-type="modal-container">
-        <SignModalMain>
-          <section> 
-            <h2>Are You a...</h2>
-            <button>Job Seeker</button>
-            <button>Employer</button>
-          </section>
-        </SignModalMain>
-      </ModalContainer>
+        <section> 
+          <h2>Are You a...</h2>
+          <button>Job Seeker</button>
+          <button>Employer</button>
+        </section>
     );
   }
 }
