+++ client/src/reducer.js
@@ -7,8 +7,9 @@ import SeekerFavorites from './components/Seeker/SeekerFavorites';
 import SeekerSettings from './components/SeekerProfile/SeekerProfile';
 import SeekerProfile from './components/SeekerProfile/SeekerProfile';
 import SignIn from './components/SignIn/SignIn';
-import SeekerSignUp from './components/SignUp/SeekerSignUp';
-import EmployerSignUp from './components/SignUp/EmployerSignUp';
+import SignUp from './components/SignUp/SignUp';
+import SeekerSignUp from './components/SignUp/Forms/SeekerSignUp';
+import EmployerSignUp from './components/SignUp/Forms/EmployerSignUp';
 
 export {
     NavBar,
@@ -20,6 +21,7 @@ export {
     SeekerSettings,
     SeekerProfile,
     SignIn,
+    SignUp,
     SeekerSignUp,
     EmployerSignUp
 }
